Added DashboardCard Moved data classes out of components package
