+++ app/src/main/java/com/amalitech/bookmeetingroom/MainActivity.kt
@@ -6,8 +6,9 @@ import androidx.activity.compose.setContent
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import com.amalitech.core_ui.components.DashBoardCard
-import com.amalitech.core_ui.components.DashboardCardItem
+import com.amalitech.admin.components.DashBoardCard
+import com.amalitech.admin.DashboardCardItem
+import com.amalitech.core_ui.R
 import com.amalitech.core_ui.theme.BookMeetingRoomTheme
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
@@ -20,7 +21,7 @@ class MainActivity : ComponentActivity() {
         setContent {
             BookMeetingRoomTheme {
                 DashBoardCard(
-                    DashboardCardItem(label = "Users", iconId = com.amalitech.core_ui.R.drawable.baseline_mail_outline_24, count = 5)
+                    DashboardCardItem(label = "Users", iconId = R.drawable.baseline_mail_outline_24, count = 5)
                 )
             }
         }
