+++ ui/admin/src/main/java/com/amalitech/admin/DashboardCardItem.kt
@@ -1,8 +1,7 @@
-package com.amalitech.core_ui.components
+package com.amalitech.admin
 
 import androidx.annotation.DrawableRes
 
-// TODO(Move this file into the admin module)
 data class DashboardCardItem(
     val label: String,
     val count: Int,
