+++ ui/admin/src/main/java/com/amalitech/admin/components/DashboardBarGraph.kt
@@ -20,6 +20,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import com.amalitech.admin.R
+import com.amalitech.admin.RoomsBookedTime
 import com.amalitech.core_ui.theme.LocalSpacing
 import com.github.mikephil.charting.charts.BarChart
 import com.github.mikephil.charting.components.XAxis
